Allow null Session.start date
When a Session is first created it does not have a start date.
A start date is first set when the presentation is 'activated'.

Remove instance variable Session in MemberRegistration.
MemberRegistration is a Session bean and belong to a pool.
If this session instance were to be used it would be random
which instance the given method would be operating on.

fixes #3